exec directory refactor
