+++ exec.spec/operator.go
