+++ exec.spec/stack.go
