+++ exec/bytecode/context.go
@@ -14,10 +14,10 @@
  limitations under the License.
 */
 
-package exec
+package bytecode
 
 import (
-	"github.com/qiniu/qlang/v6/exec/spec"
+	exec "github.com/qiniu/qlang/v6/exec.spec"
 	"github.com/qiniu/x/log"
 )
