+++ exec/bytecode/interface.go
@@ -1,9 +1,9 @@
-package exec
+package bytecode
 
 import (
 	"reflect"
 
-	exec "github.com/qiniu/qlang/v6/exec/spec"
+	exec "github.com/qiniu/qlang/v6/exec.spec"
 	"github.com/qiniu/x/log"
 )
