+++ exec/go/printer/printer.go
@@ -1358,7 +1358,7 @@ func (cfg *Config) Fprint(output io.Writer, fset *token.FileSet, node interface{
 // Fprint "pretty-prints" an AST node to output.
 // It calls Config.Fprint with default settings.
 // Note that gofmt uses tabs for indentation but spaces for alignment;
-// use format.Node (package go/format) for output that matches gofmt.
+// use format.Node (package golang/format) for output that matches gofmt.
 //
 func Fprint(output io.Writer, fset *token.FileSet, node interface{}) error {
 	return (&Config{Tabwidth: 8}).Fprint(output, fset, node)
