+++ exec/golang/testdata/const/bar.ql
